marketplace_loader/requirement.txt
azure-storage-blob==12.5.0
numpy==1.20.1
pandas==1.1.5
pyarrow==3.0.0
requests==2.23.0
urllib3==1.25.11
configparser==5.0.2
google-cloud-storage
snowflake
snowflake-connector-python
flask 
gunicorn
ZipFile38
sqlalchemy
bs4
